split user proxies into separate files
