+++ codeforlife/user/models/user/non_admin_school_teacher.py
@@ -1,0 +1,52 @@
+"""
+© Ocado Group
+Created on 05/02/2024 at 09:50:04(+00:00).
+"""
+
+import typing as t
+
+from django.db.models.query import QuerySet
+
+from ..school import School
+from .school_teacher import SchoolTeacherUser, SchoolTeacherUserManager
+from .user import User
+
+if t.TYPE_CHECKING:  # pragma: no cover
+    from django_stubs_ext.db.models import TypedModelMeta
+else:
+    TypedModelMeta = object
+
+AnyUser = t.TypeVar("AnyUser", bound=User)
+
+
+# pylint: disable-next=missing-class-docstring,too-few-public-methods
+class NonAdminSchoolTeacherUserManager(
+    SchoolTeacherUserManager["NonAdminSchoolTeacherUser"]
+):
+    def filter_users(self, queryset: QuerySet[User]):
+        return (
+            super().filter_users(queryset).filter(new_teacher__is_admin=False)
+        )
+
+
+# pylint: disable-next=too-many-ancestors
+class NonAdminSchoolTeacherUser(SchoolTeacherUser):
+    """A user that is a non-admin-teacher in a school."""
+
+    class Meta(TypedModelMeta):
+        proxy = True
+
+    objects: NonAdminSchoolTeacherUserManager = (  # type: ignore[misc]
+        NonAdminSchoolTeacherUserManager()  # type: ignore[assignment]
+    )
+
+    @property
+    def teacher(self):
+        # pylint: disable-next=import-outside-toplevel
+        from ..teacher import NonAdminSchoolTeacher, teacher_as_type
+
+        teacher = super().teacher
+
+        return (
+            teacher_as_type(teacher, NonAdminSchoolTeacher) if teacher else None
+        )
